Moved regular_polygon and random_points into a new Generators module.
