+++ spec/lib/mb/geometry/generators_spec.rb
@@ -1,0 +1,4 @@
+RSpec.describe(MB::Geometry::Generators) do
+  pending '.regular_polygon'
+  pending '.random_points'
+end
